# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Dokumentácia – Príjmy a Výdavky",
    "version": "1.0.0",
    "description": "Rozhranie pre prácu s príjmami (Incomes) a výdavkami (Receipts & Receipt Items)."
  },
  "servers": [
    { "url": "/" }
  ],
  "paths": {
...
    "/api/receipts": {
      "get": {
        "summary": "Získanie zoznamu všetkých účteniek",
        "description": "Vráti zoznam všetkých účteniek. Možno použiť voliteľné parametre na triedenie podľa dátumu alebo sumy a filtrovanie podľa mesiaca/roku.",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "schema": { "type": "string", "enum": ["issue_date", "total_amount"], "default": "issue_date" },
            "description": "Triedenie podľa poľa (issue_date alebo total_amount)"
          },
          {
            "in": "query",
            "name": "order",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" },
            "description": "Poradie triedenia – vzostupne alebo zostupne"
          },
          {
            "in": "query",
            "name": "year",
            "schema": { "type": "integer", "example": 2025 },
            "description": "Rok (použiť spolu s month)"
          },
          {
            "in": "query",
            "name": "month",
            "schema": { "type": "integer", "minimum": 1, "maximum": 12, "example": 10 },
            "description": "Mesiac (1..12) (použiť spolu s year)"
          }
        ],
        "responses": {
          "200": {
            "description": "Úspešne načítané účtenky",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                    "external_uid": "INV-2025-00012",
                    "tag": "Groceries",
                    "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                    "description": "Lidl - potraviny",
                    "issue_date": "2025-10-27",
                    "currency": "EUR",
                    "total_amount": 45.20,
                    "extra_metadata": {
                      "payment_method": "card",
                      "vat": 7.5
                    },
                    "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                    "created_at": "2025-10-27T19:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Neplatné vstupné dáta",
            "content": {
              "application/json": {
                "examples": {
                  "invalid_year_month_format": {
                    "summary": "Neplatný formát year/month",
                    "value": { "error": "Invalid year/month format" }
                  },
                  "invalid_sort": {
                    "summary": "Neplatné pole pre triedenie",
                    "value": { "error": "Invalid sort field" }
                  },
                  "missing_one_param": {
                    "summary": "Chýba year alebo month",
                    "value": { "error": "Both year and month must be provided together" }
                  },
                  "month_out_of_range": {
                    "summary": "Mesiac mimo rozsahu",
                    "value": { "error": "Month must be between 1 and 12" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Vytvorenie novej účtenky",
        "description": "Vytvorí novú účtenku (receipt) priradenú k existujúcemu používateľovi.\n\n**Povinné polia:** `user_id`, `description`\n\n**Nepovinné polia:** `issue_date`, `currency` (predvolené `EUR`), `total_amount` (predvolené `0.0`), `tag_id` (musí patriť tomu istému používateľovi), `external_uid`, `extra_metadata`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                "description": "Lidl - týždenný nákup",
                "issue_date": "2025-10-27",
                "currency": "EUR",
                "total_amount": 45.20,
                "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                "external_uid": "INV-2025-00012",
                "extra_metadata": {
                  "payment_method": "card",
                  "vat": 7.5,
                  "cashier": "Marek",
                  "location": "Bratislava"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Úspešne vytvorená účtenka",
            "content": {
              "application/json": {
                "example": {
                  "id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                  "message": "Receipt created successfully"
                }
              }
            }
          },
          "400": {
            "description": "Chybná požiadavka – chýbajú povinné polia alebo zlé formáty dát",
            "content": {
              "application/json": {
                "examples": {
                  "missing_body": {
                    "summary": "Chýbajúce telo požiadavky",
                    "value": { "error": "Missing JSON body" }
                  },
                  "validation_error": {
                    "summary": "Chyba validácie",
                    "value": {
                      "error": "Missing user_id"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/{receipt_id}": {
      "get": {
        "summary": "Získanie detailu účtenky podľa ID",
        "parameters": [
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID účtenky"
          }
        ],
        "responses": {
          "200": {
            "description": "Úspešne načítaná účtenka",
            "content": {
              "application/json": {
                "example": {
                  "id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                  "external_uid": "INV-2025-00012",
                  "tag": "Groceries",
                  "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                  "description": "Lidl - potraviny",
                  "issue_date": "2025-10-27",
                  "currency": "EUR",
                  "total_amount": 45.20,
                  "extra_metadata": {
                    "payment_method": "card",
                    "vat": 7.5
                  },
                  "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                  "created_at": "2025-10-27T19:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Účtenka nebola nájdená",
            "content": {
              "application/json": {
                "example": { "error": "Receipt not found" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Aktualizácia existujúcej účtenky",
        "description": "Aktualizuje záznam účtenky podľa ID.\n\n**Nepovinné polia:** `description`, `issue_date`, `currency`, `total_amount`, `tag_id` (null = odpojiť tag), `external_uid`, `extra_metadata`.",
        "parameters": [
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "description": "Tesco - týždenný nákup",
                "issue_date": "2025-10-29",
                "currency": "EUR",
                "total_amount": 60.50,
                "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                "external_uid": "INV-2025-00013",
                "extra_metadata": {
                  "payment_method": "cash",
                  "vat": 9.5,
                  "cashier": "Mária"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Úspešne aktualizovaná účtenka",
            "content": {
              "application/json": {
                "example": {
                  "id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                  "message": "Receipt updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Neplatné dáta alebo formát dátumu",
            "content": {
              "application/json": {
                "examples": {
                  "empty_description": {
                    "summary": "Prázdny popis",
                    "value": { "error": "Description cannot be empty" }
                  },
                  "invalid_date": {
                    "summary": "Neplatný formát dátumu",
                    "value": { "error": "Invalid issue_date format, expected YYYY-MM-DD" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Účtenka nebola nájdená",
            "content": {
              "application/json": {
                "example": { "error": "Receipt not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Odstránenie účtenky podľa ID",
        "parameters": [
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Účtenka bola úspešne odstránená",
            "content": {
              "application/json": {
                "example": { "message": "Receipt deleted successfully" }
              }
            }
          },
          "404": {
            "description": "Účtenka nebola nájdená",
            "content": {
              "application/json": {
                "example": { "error": "Receipt not found" }
              }
            }
          }
        }
      }
    },
    "/api/receipts/tags": {
      "get": {
        "summary": "Získanie všetkých expense tagov",
        "description": "Vráti zoznam tagov, ktoré sú typu EXPENSE alebo BOTH. Voliteľne je možné filtrovať podľa používateľa.",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID používateľa (voliteľné)"
          }
        ],
        "responses": {
          "200": {
            "description": "Úspešne načítané",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "tags": [
                    {
                      "id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                      "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                      "name": "Groceries",
                      "type": "EXPENSE",
                      "counter": 3
                    },
                    {
                      "id": "aa5f9e12-5b74-4d0e-aeb8-8a3e8a3be111",
                      "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                      "name": "Home",
                      "type": "BOTH",
                      "counter": 5
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Neplatné vstupné dáta",
            "content": {
              "application/json": {
                "example": { "error": "Invalid user_id format" }
              }
            }
          }
        }
      }
    },
    "/api/receipts/ekasa-items": {
      "get": {
        "summary": "Získanie eKasa položiek za konkrétny mesiac+rok (zoskupené podľa bločku)",
        "description": "Vráti všetky eKasa položky (receipt items) za zadaný mesiac a rok. Položky patriace k rovnakému bločku (účtenke) sú vrátené ako pole `items` pod jedným záznamom v `checks`.\n\n**Poznámka:** Endpoint vracia iba účtenky importované z eKasa (napr. tie, ktoré majú `external_uid`).",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "schema": { "type": "integer", "example": 2025 },
            "description": "Rok (použiť spolu s month)"
          },
          {
            "in": "query",
            "name": "month",
            "schema": { "type": "integer", "minimum": 1, "maximum": 12, "example": 10 },
            "description": "Mesiac (1..12) (použiť spolu s year)"
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID používateľa (voliteľné) – filtruje iba bločky daného používateľa"
          }
        ],
        "responses": {
          "200": {
            "description": "Úspešne načítané",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "checks": [
                    {
                      "receipt_id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                      "external_uid": "1234567890ABCDEF",
                      "issue_date": "2025-10-27",
                      "description": "Lidl - potraviny",
                      "currency": "EUR",
                      "total_amount": 45.2,
                      "tag": "Groceries",
                      "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                      "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6",
                      "items": [
                        {
                          "id": "dfcd4441-8a3a-4d62-bac9-2eb7a6dbb2e9",
                          "name": "Mlieko",
                          "quantity": 2,
                          "unit_price": 1.2,
                          "total_price": 2.4,
                          "category_id": null,
                          "extra_metadata": {
                            "vatRate": 20,
                            "itemType": "T"
                          }
                        },
                        {
                          "id": "7a05d2e9-2b9f-4f26-bb3b-6b0b6a6d2f98",
                          "name": "Chlieb",
                          "quantity": 1,
                          "unit_price": 1.5,
                          "total_price": 1.5,
                          "category_id": "a3f84b9a-42cc-4ef9-9f1d-84e7e3e2d56d",
                          "extra_metadata": null
                        }
                      ]
                    }
                  ],
                  "total_checks": 1,
                  "total_items": 2
                }
              }
            }
          },
          "400": {
            "description": "Neplatné vstupné dáta",
            "content": {
              "application/json": {
                "examples": {
                  "invalid_year_month_format": {
                    "summary": "Neplatný formát year/month",
                    "value": { "error": "Invalid year/month format" }
                  },
                  "missing_one_param": {
                    "summary": "Chýba year alebo month",
                    "value": { "error": "Both year and month must be provided together" }
                  },
                  "month_out_of_range": {
                    "summary": "Mesiac mimo rozsahu",
                    "value": { "error": "Month must be between 1 and 12" }
                  },
                  "invalid_user_id": {
                    "summary": "Neplatný formát user_id",
                    "value": { "error": "Invalid user_id format" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/import-ekasa": {
      "post": {
        "summary": "Import účtenky z eKasa systému",
        "description": "Na základe `receiptId` z eKasa API načíta údaje o účtenke a uloží ich do lokálnej databázy pod daného používateľa.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "receiptId": "1234567890ABCDEF",
                "user_id": "78754f0f-e6ed-4a10-a09a-26da4cb2a5f6"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Úspešne importovaná účtenka",
            "content": {
              "application/json": {
                "example": {
                  "message": "Receipt imported successfully",
                  "tag": "Lidl - potraviny",
                  "tag_id": "12c23b3e-8fab-4df7-95dc-8a3e3d79c9f1",
                  "receipt_id": "c0e4f248-f14d-4f53-b92c-9f5ecb8239d4",
                  "total_items": 5
                }
              }
            }
          },
          "400": {
            "description": "Chyba pri volaní eKasa API alebo validácii vstupov",
            "content": {
              "application/json": {
                "examples": {
                  "missing_fields": {
                    "summary": "Chýbajúce polia",
                    "value": {
                      "code": "bad_request",
                      "message": "Missing receiptId or user_id"
                    }
                  },
                  "ekasa_error": {
                    "summary": "Chyba eKasa / importu",
                    "value": {
                      "code": "ekasa_error",
                      "message": "Import failed: ...",
                      "details": {
                        "error": "Import failed: (sqlite3.OperationalError) ..."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}